requirements.txt
streamlit==1.52.0
torch==2.9.0
torchvision==0.24.0
Pillow==11.3.0
matplotlib==3.10.0